Add shared libs
