+++ apps/dispatcher/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@howdypix"
+}
