+++ apps/dispatcher/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@howdypix/dispatcher",
+  "version": "0.1.0",
+  "description": "Dispatch work to the workers to parse the pictures.",
+  "main": "index.js",
+  "author": "Kevin Vicrey <kevin.vicrey@nexu.org>",
+  "license": "MIT",
+  "private": true,
+  "workspaces": {
+    "nohoist": [
+      "@howdypix/typescript-config"
+    ]
+  },
+  "scripts": {
+    "start": "node build/index",
+    "dev": "yarn build && yarn watch",
+    "watch": "yarn build --skip-initial-build --watch & nodemon build/index",
+    "build": "babel src --out-dir build --extensions '.ts' --verbose",
+    "lint": "eslint src/**/*.{ts,tsx}",
+    "typecheck": "tsc --noEmit",
+    "test": "jest"
+  },
+  "dependencies": {
+    "config": "^3.2.2",
+    "express": "^4.17.1",
+    "ws": "^7.1.2"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.5.5",
+    "@howdypix/babel-preset": "^0.1.0",
+    "@howdypix/eslint-config": "^0.1.0",
+    "@howdypix/graphql-schema": "^0.1.0",
+    "@howdypix/jest-config": "^0.1.0",
+    "@howdypix/typescript-config": "^0.1.0",
+    "@types/config": "^0.0.34",
+    "@types/express": "^4.17.1",
+    "@types/jest": "^24.0.18",
+    "@types/ws": "^6.0.3",
+    "eslint": " ^6.3.0",
+    "jest": "^24.9.0",
+    "nodemon": "^1.19.2",
+    "typescript": "^3.6.2"
+  }
+}
