+++ apps/server/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require("@howdypix/jest-config");
