+++ apps/server/package.json
@@ -1,11 +1,16 @@
 {
-  "name": "@howdypix/dispatcher",
+  "name": "@howdypix/server",
   "version": "0.1.0",
-  "description": "Dispatch work to the workers to parse the pictures.",
+  "description": "Server for the Photo Gallery",
   "main": "index.js",
   "author": "Kevin Vicrey <kevin.vicrey@nexu.org>",
   "license": "MIT",
   "private": true,
+  "workspaces": {
+    "nohoist": [
+      "@howdypix/typescript-config"
+    ]
+  },
   "scripts": {
     "start": "node build/index",
     "dev": "yarn build --watch & nodemon build/index",
@@ -14,27 +19,26 @@
     "typecheck": "tsc --noEmit",
     "test": "jest"
   },
-  "dependencies": {
-    "express": "^4.17.1",
-    "ws": "^7.1.2"
-  },
   "devDependencies": {
     "@babel/cli": "^7.5.5",
     "@babel/core": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-typescript": "^7.3.3",
+    "@howdypix/babel-preset": "^0.1.0",
+    "@howdypix/eslint-config": "^0.1.0",
     "@howdypix/graphql-schema": "^0.1.0",
-    "@types/express": "^4.17.1",
+    "@howdypix/jest-config": "^0.1.0",
+    "@howdypix/typescript-config": "^0.1.0",
     "@types/jest": "^24.0.18",
-    "@types/ws": "^6.0.3",
-    "@typescript-eslint/parser": "^2.0.0",
-    "babel-jest": "^24.9.0",
     "eslint": " ^6.3.0",
-    "eslint-config-node": "^4.0.0",
-    "eslint-config-prettier": "^6.1.0",
-    "eslint-plugin-import": "^2.18.2",
     "jest": "^24.9.0",
     "nodemon": "^1.19.2",
     "typescript": "^3.6.2"
+  },
+  "dependencies": {
+    "apollo-server": "^2.9.3",
+    "express": "^4.17.1",
+    "globby": "^10.0.1",
+    "graphql": "^14.5.5",
+    "nexus": "^0.11.7",
+    "ws": "^7.1.2"
   }
 }
