+++ apps/server/src/schema/hello.ts
