+++ libs/babel-preset/node.js
@@ -1,0 +1,15 @@
+module.exports = function() {
+  return {
+    presets: [
+      [
+        "@babel/preset-env",
+        {
+          targets: {
+            node: "current"
+          }
+        }
+      ],
+      "@babel/preset-typescript"
+    ]
+  };
+};
