+++ libs/babel-preset/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@howdypix/babel-preset",
+  "version": "0.1.0",
+  "license": "MIT",
+  "devDependencies": {
+    "@babel/core": "^7.0.0"
+  },
+  "dependencies": {
+    "@babel/preset-env": "^7.6.0",
+    "@babel/preset-typescript": "^7.6.0",
+    "babel-jest": "^24.9.0"
+  }
+}
