+++ libs/graphql-schema/scripts/graphql2json.js
