+++ libs/jest-config/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@howdypix/jest-config",
+  "version": "0.1.0",
+  "license": "MIT",
+  "dependencies": {}
+}
