+++ libs/typescript-config/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@howdypix/typescript-config",
+  "version": "0.1.0",
+  "license": "MIT",
+  "main": "tsconfig.json",
+  "dependencies": {}
+}
