+++ libs/typescript-config/tsconfig.json
@@ -6,7 +6,5 @@
     "outDir": "build",
     "strict": true,
     "esModuleInterop": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules", "**/*.test.ts"]
+  }
 }
