+++ package.json
@@ -3,6 +3,12 @@
   "version": "0.1.0",
   "description": "Open-Source Photo Gallery.",
   "main": "index.js",
+  "workspaces": {
+    "packages": [
+      "apps/*",
+      "libs/*"
+    ]
+  },
   "repository": "git@github.com:vickev/howdypix.git",
   "author": "Kevin Vicrey <kevin.vicrey@nexu.org>",
   "license": "MIT",
