--- packages/server/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@howdypix/server",
-  "version": "0.1.0",
-  "description": "Server for the Photo Gallery",
-  "main": "index.js",
-  "author": "Kevin Vicrey <kevin.vicrey@nexu.org>",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "start": "node build/index",
-    "dev": "yarn build --watch & nodemon build/index",
-    "build": "babel src --out-dir build --extensions '.ts' --verbose",
-    "lint": "eslint src/**/*.{ts,tsx}",
-    "typecheck": "tsc --noEmit",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-typescript": "^7.3.3",
-    "@howdypix/graphql-schema": "^0.1.0",
-    "@types/jest": "^24.0.18",
-    "@typescript-eslint/parser": "^2.0.0",
-    "babel-jest": "^24.9.0",
-    "eslint": " ^6.3.0",
-    "eslint-config-node": "^4.0.0",
-    "eslint-config-prettier": "^6.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "jest": "^24.9.0",
-    "nodemon": "^1.19.2",
-    "typescript": "^3.6.2"
-  },
-  "dependencies": {
-    "apollo-server": "^2.9.3",
-    "express": "^4.17.1",
-    "globby": "^10.0.1",
-    "graphql": "^14.5.5",
-    "nexus": "^0.11.7",
-    "ws": "^7.1.2"
-  }
-}
